+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.0] - 2026-01-27
+
 ### Added
 
 - **HATEOAS-style Response Actions** - Add `_actions` to responses (#98, #99):